+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.0-rc.9",
+  sdkVersion: "3.14.0-rc.10",
   genVersion: "2.721.3",
-  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.9 2.721.3 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.10 2.721.3 2.0 @orq-ai/node",
 } as const;