+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -751,7 +751,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6W9YKEWVF0G1JKN17AK2TYZ"),
+  _id: z.string().default("01K6WN6YKJ85G362TV3D3XG1BC"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -795,7 +795,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K6W9YKEWVF0G1JKN17AK2TYZ"),
+  id: z.string().default("01K6WN6YKJ85G362TV3D3XG1BC"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,