+++ packages/orq-rc/RELEASES.md
@@ -4099,3 +4099,13 @@ Based on:
 - [typescript v3.14.0-rc.9] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.9 - packages/orq-rc
+
+## 2025-10-06 11:42:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.2 (2.721.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.10] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.10 - packages/orq-rc